+++ server/checkmate/game/wsconsumers/Game.py
@@ -1,0 +1,13 @@
+from channels.generic.websocket import WebsocketConsumer
+
+class GameConsumer(WebsocketConsumer):
+    def connect(self):
+        self.accept()
+        
+    def disconnect(self, close_code):
+        pass
+    
+    def receive(self, text_data):
+        self.send(text_data=text_data)
+        
+
